seperate out dis, io
